+++ packages/kit/types/internal.d.ts
@@ -1,20 +1,41 @@
 import { OutputAsset, OutputChunk } from 'rollup';
-import { ValidatedConfig } from './config';
-import { InternalApp, SSRManifest } from './app';
-import { Fallthrough, RequestHandler, ShadowRequestHandler } from './endpoint';
-import { Either } from './helper';
-import { ExternalFetch, GetSession, Handle, HandleError, RequestEvent } from './hooks';
-import { Load } from './page';
+import {
+	SSRManifest,
+	ValidatedConfig,
+	RequestHandler,
+	Load,
+	ExternalFetch,
+	GetSession,
+	Handle,
+	HandleError,
+	RequestEvent,
+	App,
+	RequestOptions,
+	PrerenderErrorHandler
+} from './index';
 
-export interface PrerenderDependency {
-	response: Response;
-	body: null | string | Uint8Array;
-}
+export interface AdapterEntry {
+	/**
+	 * A string that uniquely identifies an HTTP service (e.g. serverless function) and is used for deduplication.
+	 * For example, `/foo/a-[b]` and `/foo/[c]` are different routes, but would both
+	 * be represented in a Netlify _redirects file as `/foo/:param`, so they share an ID
+	 */
+	id: string;
 
-export interface PrerenderOptions {
-	fallback?: string;
-	all: boolean;
-	dependencies: Map<string, PrerenderDependency>;
+	/**
+	 * A function that compares the candidate route with the current route to determine
+	 * if it should be treated as a fallback for the current route. For example, `/foo/[c]`
+	 * is a fallback for `/foo/a-[b]`, and `/[...catchall]` is a fallback for all routes
+	 */
+	filter: (route: RouteDefinition) => boolean;
+
+	/**
+	 * A function that is invoked once the entry has been created. This is where you
+	 * should write the function to the filesystem and generate redirect manifests.
+	 */
+	complete: (entry: {
+		generateManifest: (opts: { relativePath: string; format?: 'esm' | 'cjs' }) => string;
+	}) => void;
 }
 
 export interface AppModule {
@@ -31,6 +52,84 @@ export interface AppModule {
 	}): void;
 }
 
+export interface Asset {
+	file: string;
+	size: number;
+	type: string | null;
+}
+
+export type Body = JSONValue | Uint8Array | ReadableStream | import('stream').Readable;
+
+export interface BuildData {
+	app_dir: string;
+	manifest_data: ManifestData;
+	service_worker: string | null;
+	client: {
+		assets: OutputAsset[];
+		chunks: OutputChunk[];
+		entry: {
+			file: string;
+			js: string[];
+			css: string[];
+		};
+		vite_manifest: import('vite').Manifest;
+	};
+	server: {
+		chunks: OutputChunk[];
+		methods: Record<string, HttpMethod[]>;
+		vite_manifest: import('vite').Manifest;
+	};
+	static: string[];
+	entries: string[];
+}
+
+export type CSRComponent = any; // TODO
+
+export type CSRComponentLoader = () => Promise<CSRComponent>;
+
+export type CSRRoute = [RegExp, CSRComponentLoader[], CSRComponentLoader[], GetParams?, HasShadow?];
+
+export type Either<T, U> = Only<T, U> | Only<U, T>;
+
+export interface EndpointData {
+	type: 'endpoint';
+	key: string;
+	segments: RouteSegment[];
+	pattern: RegExp;
+	params: string[];
+	file: string;
+}
+
+export interface Fallthrough {
+	fallthrough: true;
+}
+
+export type GetParams = (match: RegExpExecArray) => Record<string, string>;
+
+type HasShadow = 1;
+
+export interface Hooks {
+	externalFetch: ExternalFetch;
+	getSession: GetSession;
+	handle: Handle;
+	handleError: HandleError;
+}
+
+export type HttpMethod = 'get' | 'head' | 'post' | 'put' | 'delete' | 'patch';
+
+export class InternalApp extends App {
+	render(
+		request: Request,
+		options?: RequestOptions & {
+			prerender?: PrerenderOptions;
+		}
+	): Promise<Response>;
+}
+
+export type JSONObject = { [key: string]: JSONValue };
+
+export type JSONValue = string | number | boolean | null | ToJSON | JSONValue[] | JSONObject;
+
 export interface Logger {
 	(msg: string): void;
 	success(msg: string): void;
@@ -38,6 +137,117 @@ export interface Logger {
 	warn(msg: string): void;
 	minor(msg: string): void;
 	info(msg: string): void;
+}
+
+export interface ManifestData {
+	assets: Asset[];
+	layout: string;
+	error: string;
+	components: string[];
+	routes: RouteData[];
+}
+
+export type MaybePromise<T> = T | Promise<T>;
+
+export interface MethodOverride {
+	parameter: string;
+	allowed: string[];
+}
+
+export type NormalizedLoadOutput = Either<
+	{
+		status: number;
+		error?: Error;
+		redirect?: string;
+		props?: Record<string, any> | Promise<Record<string, any>>;
+		stuff?: Record<string, any>;
+		maxage?: number;
+	},
+	Fallthrough
+>;
+
+type Only<T, U> = { [P in keyof T]: T[P] } & { [P in Exclude<keyof U, keyof T>]?: never };
+
+export interface PageData {
+	type: 'page';
+	key: string;
+	shadow: string | null;
+	segments: RouteSegment[];
+	pattern: RegExp;
+	params: string[];
+	path: string;
+	a: string[];
+	b: string[];
+}
+
+export interface PrerenderDependency {
+	response: Response;
+	body: null | string | Uint8Array;
+}
+
+export type PrerenderOnErrorValue = 'fail' | 'continue' | PrerenderErrorHandler;
+
+export interface PrerenderOptions {
+	fallback?: string;
+	all: boolean;
+	dependencies: Map<string, PrerenderDependency>;
+}
+
+export type RecursiveRequired<T> = {
+	// Recursive implementation of TypeScript's Required utility type.
+	// Will recursively continue until it reaches primitive or union
+	// with a Function in it, except those commented below
+	[K in keyof T]-?: Extract<T[K], Function> extends never // If it does not have a Function type
+		? RecursiveRequired<T[K]> // recursively continue through.
+		: K extends 'vite' // If it reaches the 'vite' key
+		? Extract<T[K], Function> // only take the Function type.
+		: T[K]; // Use the exact type for everything else
+};
+
+export interface RequiredResolveOptions {
+	ssr: boolean;
+	transformPage: ({ html }: { html: string }) => string;
+}
+
+export interface Respond {
+	(request: Request, options: SSROptions, state?: SSRState): Promise<Response>;
+}
+
+/** `string[]` is only for set-cookie, everything else must be type of `string` */
+export type ResponseHeaders = Record<string, string | number | string[]>;
+
+export type RouteData = PageData | EndpointData;
+
+export interface RouteDefinition {
+	type: 'page' | 'endpoint';
+	pattern: RegExp;
+	segments: RouteSegment[];
+	methods: HttpMethod[];
+}
+
+export interface RouteSegment {
+	content: string;
+	dynamic: boolean;
+	rest: boolean;
+}
+
+export interface ShadowEndpointOutput<Output extends JSONObject = JSONObject> {
+	status?: number;
+	headers?: Partial<ResponseHeaders>;
+	body?: Output;
+}
+
+export interface ShadowRequestHandler<Output extends JSONObject = JSONObject> {
+	(event: RequestEvent): MaybePromise<Either<ShadowEndpointOutput<Output>, Fallthrough>>;
+}
+
+export interface ShadowData {
+	fallthrough?: boolean;
+	status?: number;
+	error?: Error;
+	redirect?: string;
+	cookies?: string[];
+	body?: JSONObject;
 }
 
 export interface SSRComponent {
@@ -59,37 +269,6 @@ export interface SSRComponent {
 
 export type SSRComponentLoader = () => Promise<SSRComponent>;
 
-export type CSRComponent = any; // TODO
-
-export type CSRComponentLoader = () => Promise<CSRComponent>;
-
-export interface SSRPagePart {
-	id: string;
-	load: SSRComponentLoader;
-}
-
-export type GetParams = (match: RegExpExecArray) => Record<string, string>;
-
-export interface SSRPage {
-	type: 'page';
-	pattern: RegExp;
-	params: GetParams;
-	shadow:
-		| null
-		| (() => Promise<{
-				[method: string]: ShadowRequestHandler;
-		  }>);
-	/**
-	 * plan a is to render 1 or more layout components followed by a leaf component.
-	 */
-	a: number[];
-	/**
-	 * plan b — if one of them components fails in `load` we backtrack until we find
-	 * the nearest error component.
-	 */
-	b: number[];
-}
-
 export interface SSREndpoint {
 	type: 'endpoint';
 	pattern: RegExp;
@@ -97,20 +276,6 @@ export interface SSREndpoint {
 	load(): Promise<{
 		[method: string]: RequestHandler;
 	}>;
-}
-
-export type SSRRoute = SSREndpoint | SSRPage;
-
-type HasShadow = 1;
-export type CSRRoute = [RegExp, CSRComponentLoader[], CSRComponentLoader[], GetParams?, HasShadow?];
-
-export type SSRNodeLoader = () => Promise<SSRNode>;
-
-export interface Hooks {
-	externalFetch: ExternalFetch;
-	getSession: GetSession;
-	handle: Handle;
-	handleError: HandleError;
 }
 
 export interface SSRNode {
@@ -124,6 +289,8 @@ export interface SSRNode {
 	/** inlined styles */
 	styles?: Record<string, string>;
 }
+
+export type SSRNodeLoader = () => Promise<SSRNode>;
 
 export interface SSROptions {
 	amp: boolean;
@@ -161,6 +328,33 @@ export interface SSROptions {
 	trailing_slash: TrailingSlash;
 }
 
+export interface SSRPage {
+	type: 'page';
+	pattern: RegExp;
+	params: GetParams;
+	shadow:
+		| null
+		| (() => Promise<{
+				[method: string]: ShadowRequestHandler;
+		  }>);
+	/**
+	 * plan a is to render 1 or more layout components followed by a leaf component.
+	 */
+	a: number[];
+	/**
+	 * plan b — if one of them components fails in `load` we backtrack until we find
+	 * the nearest error component.
+	 */
+	b: number[];
+}
+
+export interface SSRPagePart {
+	id: string;
+	load: SSRComponentLoader;
+}
+
+export type SSRRoute = SSREndpoint | SSRPage;
+
 export interface SSRState {
 	fetched?: string;
 	initiator?: SSRPage | null;
@@ -169,92 +363,10 @@ export interface SSRState {
 	fallback?: string;
 }
 
-export interface Asset {
-	file: string;
-	size: number;
-	type: string | null;
-}
+export type StrictBody = string | Uint8Array;
 
-export interface RouteSegment {
-	content: string;
-	dynamic: boolean;
-	rest: boolean;
-}
-
-export type HttpMethod = 'get' | 'head' | 'post' | 'put' | 'delete' | 'patch';
-
-export interface PageData {
-	type: 'page';
-	key: string;
-	shadow: string | null;
-	segments: RouteSegment[];
-	pattern: RegExp;
-	params: string[];
-	path: string;
-	a: string[];
-	b: string[];
-}
-
-export interface EndpointData {
-	type: 'endpoint';
-	key: string;
-	segments: RouteSegment[];
-	pattern: RegExp;
-	params: string[];
-	file: string;
-}
-
-export type RouteData = PageData | EndpointData;
-
-export interface ManifestData {
-	assets: Asset[];
-	layout: string;
-	error: string;
-	components: string[];
-	routes: RouteData[];
-}
-
-export interface BuildData {
-	app_dir: string;
-	manifest_data: ManifestData;
-	service_worker: string | null;
-	client: {
-		assets: OutputAsset[];
-		chunks: OutputChunk[];
-		entry: {
-			file: string;
-			js: string[];
-			css: string[];
-		};
-		vite_manifest: import('vite').Manifest;
-	};
-	server: {
-		chunks: OutputChunk[];
-		methods: Record<string, HttpMethod[]>;
-		vite_manifest: import('vite').Manifest;
-	};
-	static: string[];
-	entries: string[];
-}
-
-export type NormalizedLoadOutput = Either<
-	{
-		status: number;
-		error?: Error;
-		redirect?: string;
-		props?: Record<string, any> | Promise<Record<string, any>>;
-		stuff?: Record<string, any>;
-		maxage?: number;
-	},
-	Fallthrough
->;
+type ToJSON = { toJSON(...args: any[]): Exclude<JSONValue, ToJSON> };
 
 export type TrailingSlash = 'never' | 'always' | 'ignore';
-export interface MethodOverride {
-	parameter: string;
-	allowed: string[];
-}
 
-export interface Respond {
-	(request: Request, options: SSROptions, state?: SSRState): Promise<Response>;
-}
+export * from './index';
